+++ src/main/java/com/devbliss/doctest/items/ResponseDocItem.java
@@ -1,38 +1,35 @@
 package com.devbliss.doctest.items;
 
+import java.util.Map;
+
+import de.devbliss.apitester.ApiResponse;
+
+
 public class ResponseDocItem implements DocItem {
 
-    private int responseCode;
-    private JsonDocItem payload;
+    private final int responseCode;
+    private final JsonDocItem payload;
+    private final Map<String, String> headers;
 
-    public ResponseDocItem(int responseCode, JsonDocItem payload) {
-        this.responseCode = responseCode;
-        this.payload = payload;
-    }
-
-    public ResponseDocItem(int responseCode, String json) {
-        this.responseCode = responseCode;
-        this.payload = new JsonDocItem(json);
+    public ResponseDocItem(ApiResponse response, Map<String, String> headers) {
+        this.responseCode = response.httpStatus;
+        this.payload = new JsonDocItem(response.payload);
+        this.headers = headers;
     }
 
     public int getResponseCode() {
         return responseCode;
     }
 
-    public void setResponseCode(int responseCode) {
-        this.responseCode = responseCode;
-    }
-
     public JsonDocItem getPayload() {
         return payload;
-    }
-
-    public void setPayload(JsonDocItem payload) {
-        this.payload = payload;
     }
 
     public String getItemName() {
         return "response";
     }
 
+    public Map<String, String> getHeaders() {
+        return this.headers;
+    }
 }