+++ src/test/java/com/devbliss/doctest/machine/DocTestMachineImplUnitTest.java
@@ -9,7 +9,10 @@ import static org.mockito.Mockito.when;
 
 import java.net.URI;
 import java.net.URISyntaxException;
+import java.util.ArrayList;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 
 import org.junit.Before;
 import org.junit.Test;
@@ -28,10 +31,12 @@ import com.devbliss.doctest.items.ResponseDocItem;
 import com.devbliss.doctest.items.SectionDocItem;
 import com.devbliss.doctest.items.TextDocItem;
 import com.devbliss.doctest.renderer.ReportRenderer;
+import com.devbliss.doctest.utils.HeadersHelper;
 import com.devbliss.doctest.utils.JSONHelper;
 import com.devbliss.doctest.utils.UriHelper;
 
-import de.devbliss.apitester.ApiTest.HTTP_REQUEST;
+import de.devbliss.apitester.ApiRequest;
+import de.devbliss.apitester.ApiResponse;
 
 /**
  * Unit test for {@link DocTestMachineImpl}
@@ -48,6 +53,7 @@ public class DocTestMachineImplUnitTest {
     private static final String JSON_VALID = "{'abc':'a'}";
     private static final String JSON_INVALID = "invalid";
     private static final String INTRODUCTION = "";
+    private static final String HTTP_METHOD = "get";
 
     @Mock
     private ReportRenderer renderer;
@@ -57,23 +63,33 @@ public class DocTestMachineImplUnitTest {
     private List<DocItem> listItem;
     @Mock
     private UriHelper uriHelper;
+    @Mock
+    private HeadersHelper headersHelper;
 
     @Captor
     private ArgumentCaptor<List<DocItem>> listItemCaptor;
 
+    private ApiRequest apiRequest;
+    private ApiResponse apiResponse;
     private DocTestMachineImpl machine;
-    private final HTTP_REQUEST httpRequest = HTTP_REQUEST.GET;
-    private String uriString;
+    private final String uriString = "www.example.de:91/ressource";
     private URI uri;
+    private List<String> headersToShow;
+    private Map<String, String> headers;
 
     @Before
     public void setUp() throws URISyntaxException {
         uri = new URI("");
+        headersToShow = new ArrayList<String>();
+        headers = new HashMap<String, String>();
+
+        apiRequest = new ApiRequest(uri, HTTP_METHOD, headers);
+        apiResponse = new ApiResponse(RESPONSE_CODE, "", JSON_VALID, headers);
 
         when(jsonHelper.isJsonValid(JSON_VALID)).thenReturn(true);
         when(jsonHelper.prettyPrintJson(JSON_VALID)).thenReturn(JSON_VALID);
         when(uriHelper.uriToString(uri)).thenReturn(uriString);
-        machine = spy(new DocTestMachineImpl(renderer, jsonHelper, uriHelper));
+        machine = spy(new DocTestMachineImpl(renderer, jsonHelper, uriHelper, headersHelper));
     }
 
     @Test
@@ -130,7 +146,7 @@ public class DocTestMachineImplUnitTest {
     @Test
     public void addResponseItem() throws Exception {
         machine.beginDoctest(FILE_NAME, INTRODUCTION);
-        machine.sayResponse(RESPONSE_CODE, JSON_VALID);
+        machine.sayResponse(apiResponse, headersToShow);
         machine.endDocTest();
 
         verify(renderer).render(listItemCaptor.capture(), eq(FILE_NAME), eq(INTRODUCTION));
@@ -145,7 +161,7 @@ public class DocTestMachineImplUnitTest {
     @Test
     public void addRequestItem() throws Exception {
         machine.beginDoctest(FILE_NAME, INTRODUCTION);
-        machine.sayRequest(uri, JSON_VALID, httpRequest);
+        machine.sayRequest(apiRequest, JSON_VALID, headersToShow);
         machine.endDocTest();
 
         verify(renderer).render(listItemCaptor.capture(), eq(FILE_NAME), eq(INTRODUCTION));
@@ -154,14 +170,14 @@ public class DocTestMachineImplUnitTest {
         assertEquals(1, listItems.size());
         assertTrue(listItems.get(0) instanceof RequestDocItem);
         assertEquals(JSON_VALID, ((RequestDocItem) listItems.get(0)).getPayload().getExpected());
-        assertEquals(httpRequest, ((RequestDocItem) listItems.get(0)).getHttp());
+        assertEquals(HTTP_METHOD.toUpperCase(), ((RequestDocItem) listItems.get(0)).getHttp());
         assertEquals(uriString, ((RequestDocItem) listItems.get(0)).getUri());
     }
 
     @Test
     public void addUploadRequestItem() throws Exception {
         machine.beginDoctest(FILE_NAME, INTRODUCTION);
-        machine.sayUploadRequest(uri, httpRequest, "file", "fileBody", 10l, "");
+        machine.sayUploadRequest(apiRequest, "file", "fileBody", 10l, "", headersToShow);
         machine.endDocTest();
 
         verify(renderer).render(listItemCaptor.capture(), eq(FILE_NAME), eq(INTRODUCTION));
@@ -169,7 +185,7 @@ public class DocTestMachineImplUnitTest {
         List<DocItem> listItems = listItemCaptor.getValue();
         assertEquals(1, listItems.size());
         assertTrue(listItems.get(0) instanceof RequestUploadDocItem);
-        assertEquals(httpRequest, ((RequestUploadDocItem) listItems.get(0)).getHttp());
+        assertEquals(HTTP_METHOD.toUpperCase(), ((RequestUploadDocItem) listItems.get(0)).getHttp());
         assertEquals(uriString, ((RequestUploadDocItem) listItems.get(0)).getUri());
         assertEquals("file", ((RequestUploadDocItem) listItems.get(0)).getFileName());
         assertEquals("fileBody", ((RequestUploadDocItem) listItems.get(0)).getFileBody());
@@ -177,8 +193,9 @@ public class DocTestMachineImplUnitTest {
 
     @Test
     public void addUploadRequestItemUriIsNull() throws Exception {
+        apiRequest = new ApiRequest(null, HTTP_METHOD, headers);
         machine.beginDoctest(FILE_NAME, INTRODUCTION);
-        machine.sayUploadRequest(null, httpRequest, "file", "fileBody", 10l, "");
+        machine.sayUploadRequest(apiRequest, "file", "fileBody", 10l, "", headersToShow);
         machine.endDocTest();
 
         verify(renderer).render(listItemCaptor.capture(), eq(FILE_NAME), eq(INTRODUCTION));
@@ -189,8 +206,10 @@ public class DocTestMachineImplUnitTest {
 
     @Test
     public void addRequestItemWihtoutUri() throws Exception {
+        apiRequest = new ApiRequest(null, HTTP_METHOD, headers);
+
         machine.beginDoctest(FILE_NAME, INTRODUCTION);
-        machine.sayRequest(null, JSON_VALID, httpRequest);
+        machine.sayRequest(apiRequest, JSON_VALID, headersToShow);
         machine.endDocTest();
 
         verify(renderer).render(listItemCaptor.capture(), eq(FILE_NAME), eq(INTRODUCTION));
@@ -202,7 +221,7 @@ public class DocTestMachineImplUnitTest {
     @Test
     public void addRequestItemWihtInvalidJson() throws Exception {
         machine.beginDoctest(FILE_NAME, INTRODUCTION);
-        machine.sayRequest(uri, JSON_INVALID, httpRequest);
+        machine.sayRequest(apiRequest, JSON_INVALID, headersToShow);
         machine.endDocTest();
 
         verify(renderer).render(listItemCaptor.capture(), eq(FILE_NAME), eq(INTRODUCTION));
@@ -211,7 +230,7 @@ public class DocTestMachineImplUnitTest {
         assertEquals(1, listItems.size());
         assertTrue(listItems.get(0) instanceof RequestDocItem);
         assertEquals(JSON_INVALID, ((RequestDocItem) listItems.get(0)).getPayload().getExpected());
-        assertEquals(httpRequest, ((RequestDocItem) listItems.get(0)).getHttp());
+        assertEquals(HTTP_METHOD.toUpperCase(), ((RequestDocItem) listItems.get(0)).getHttp());
         assertEquals(uriString, ((RequestDocItem) listItems.get(0)).getUri());
     }
 
@@ -235,8 +254,8 @@ public class DocTestMachineImplUnitTest {
         machine.say(TEXT + "_first");
         machine.sayPreformatted(JSON_VALID);
         machine.say(TEXT + "_second");
-        machine.sayRequest(uri, JSON_VALID, httpRequest);
-        machine.sayResponse(RESPONSE_CODE, JSON_VALID);
+        machine.sayRequest(apiRequest, JSON_VALID, headersToShow);
+        machine.sayResponse(apiResponse, headersToShow);
         machine.sayVerify(TEXT);
         machine.sayNextSectionTitle(TEXT);
         machine.endDocTest();